+++ features/dashboard/dashboard-list/components/list-button/dashboard-list-add-button.tsx
@@ -1,6 +1,6 @@
 import Image from "next/image"
 import { useToggle } from "@common/hooks"
-import { CreateDashboardModal } from "@dashboard/modal"
+import { CreateDashboardModal } from "@dashboard/components/modal"
 import AddVioletIcon from "@/public/images/icons/add-violet-icon.svg"
 import classes from "./dashboard-list-add-button.module.css"
