+++ shared/@dashboard/components/dashboard-header/header-button/dashboard-header-invite-button.tsx
@@ -1,5 +1,4 @@
 import React from "react"
-import AddIcon from "@/public/images/icons/add-icon.svg"
 import { useToggle } from "@common/hooks"
 import { Button } from "@common/components/ui/button"
 
@@ -8,7 +7,7 @@ export default function DashboardHeaderInviteButton() {
   const style = { padding: "10px 16px" }
   return (
     <>
-      <Button iconUrl={AddIcon} buttonStyle="outline" customStyle={style} onClick={handleOpenToggle}>
+      <Button iconUrl="/images/icons/add-icon.svg" buttonStyle="outline" customStyle={style} onClick={handleOpenToggle}>
         초대하기
       </Button>
     </>
