+++ shared/@dashboard/components/dashboard-header/header/dashboard-header.tsx
@@ -2,7 +2,7 @@ import { PropsWithChildren } from "react"
 import Link from "next/link"
 import { useRouter } from "next/router"
 import { useAuthStore, useToggle } from "@common/hooks"
-import { DashboardHeaderDetail, DropdownHeaderProfile } from "@dashboard/dashboard-header"
+import { DashboardHeaderDetail, DropdownHeaderProfile } from "@dashboard/components/dashboard-header"
 import Avatar from "@common/components/ui/avatar"
 import type { Member } from "@common/types"
 import classes from "./dashboard-header.module.css"
