+++ shared/@dashboard/components/dashboard-layout/index.ts
