+++ shared/@dashboard/components/modal/index.ts
